move partition key range to core/definitions
